fastapi_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pyodbc==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
python-multipart==0.0.6
aiosmtplib==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0
cachetools==5.3.2
